backend/etl-pipeline/query-api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.9
python-dotenv==1.0.0
greenlet==3.0.3